[Refactor] pages, components 분리
