+++ src/main.hash.js
@@ -1,8 +1,8 @@
 import userStore from "./store/userStore";
-import HomePage from "./components/HomePage";
-import ProfilePage from "./components/ProfilePage";
-import NotFoundPage from "./components/NotFoundPage";
-import LoginPage from "./components/LoginPage";
+import Home from "./pages/Home";
+import Profile from "./pages/Profile";
+import NotFound from "./pages/NotFound";
+import Login from "./pages/Login";
 
 const loginAction = (e) => {
   e.preventDefault();
@@ -30,20 +30,20 @@ const profileUpdateAction = () => {
 };
 
 const routes = {
-  "/": () => HomePage(),
+  "/": () => Home(),
   "/login": () => {
     if (userStore.isLoggedIn()) {
       hashNavigate("/");
-      return HomePage();
+      return Home();
     }
-    return LoginPage();
+    return Login();
   },
   "/profile": () => {
     if (!userStore.isLoggedIn()) {
       hashNavigate("/login");
-      return LoginPage();
+      return Login();
     }
-    return ProfilePage();
+    return Profile();
   },
 };
 
@@ -53,7 +53,7 @@ const hashNavigate = (path) => {
 
 const App = () => {
   const path = window.location.hash.slice(1);
-  const PageContent = routes[path] || (() => NotFoundPage());
+  const PageContent = routes[path] || (() => NotFound());
   return PageContent();
 };
