+++ src/pages/Login.js
@@ -1,4 +1,4 @@
-const LoginPage = () => `
+const Login = () => `
   <div class="bg-gray-100 flex items-center justify-center min-h-screen">
     <div class="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
       <h1 class="text-2xl font-bold text-center text-blue-600 mb-8">항해플러스</h1>
@@ -22,4 +22,4 @@ const LoginPage = () => `
   </div>
 `;
 
-export default LoginPage;
+export default Login;
