+++ src/pages/NotFound.js
@@ -1,4 +1,4 @@
-const NotFoundPage = () => `
+const NotFound = () => `
   <main class="bg-gray-100 flex items-center justify-center min-h-screen" role="main">
     <div class="bg-white p-8 rounded-lg shadow-md w-full text-center" style="max-width: 480px">
       <h1 class="text-2xl font-bold text-blue-600 mb-4">항해플러스</h1>
@@ -14,4 +14,4 @@ const NotFoundPage = () => `
   </main>
 `;
 
-export default NotFoundPage;
+export default NotFound;
