+++ src/pages/Profile.js
@@ -1,8 +1,8 @@
 import userStore from "../store/userStore";
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 
-const ProfilePage = () => {
+const Profile = () => {
   const user = userStore.getUser();
 
   return `
@@ -73,4 +73,4 @@ const ProfilePage = () => {
   `;
 };
 
-export default ProfilePage;
+export default Profile;
